Add support for AWS Elasticsearch
AWS Elasticsearch requires signed requests for all API calls. The
aws-sdk-js library does not have support built in for Elasticsearch
domains, however it has a mechanism for defining services and APIs.

This defines an `ElasticsearchDomain` service with a `bulkInsert`
method. The aws-sdk library takes care of signing requests to all
elasticsearch endpoints.